feat: mode에 따른 분기처리하는 Interactionlayer
